+++ packages/atomic/src/components/search/atomic-did-you-mean/atomic-did-you-mean.stories.tsx
@@ -1,0 +1,20 @@
+import defaultStory from '../../../../.storybook/default-story';
+
+const {defaultModuleExport, exportedStory} = defaultStory(
+  'Atomic/DidYouMean',
+  'atomic-did-you-mean',
+  {},
+  {
+    engineConfig: {
+      preprocessRequest: (r) => {
+        const parsed = JSON.parse(r.body as string);
+        parsed.q = 'testt';
+        r.body = JSON.stringify(parsed);
+        return r;
+      },
+    },
+  }
+);
+
+export default defaultModuleExport;
+export const DefaultDidYouMean = exportedStory;