+++ packages/atomic/src/components/search/result-template-components/atomic-result-number/atomic-result-number.stories.tsx
@@ -1,0 +1,25 @@
+import {html} from 'lit-html';
+import defaultResultComponentStory from '../../../../../.storybook/default-result-component-story';
+
+const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
+  'Atomic/ResultList/ResultNumber',
+  'atomic-result-number',
+  {field: 'size'},
+  {
+    engineConfig: {
+      preprocessRequest: (r) => {
+        const parsed = JSON.parse(r.body as string);
+        parsed.fieldsToInclude = [...parsed.fieldsToInclude, 'size'];
+        r.body = JSON.stringify(parsed);
+        return r;
+      },
+    },
+    additionalChildMarkup: () => html`
+      <span style="font-weight: bold; margin-right: 0.25rem;">File size:</span>
+      <atomic-format-unit unit="byte" unit-display="long"></atomic-format-unit>
+    `,
+  }
+);
+
+export default defaultModuleExport;
+export const DefaultResultNumber = exportedStory;