+++ packages/headless/src/controllers/facets/category-facet/headless-category-facet.test.ts
@@ -39,7 +39,6 @@ import {
   configuration,
   search,
 } from '../../../app/reducers';
-import {defaultFacetSearchOptions} from '../../../features/facets/facet-search-set/facet-search-reducer-helpers';
 
 describe('category facet', () => {
   const facetId = '1';
@@ -97,7 +96,6 @@ describe('category facet', () => {
   it('registers a category facet with the passed options and default optional parameters', () => {
     const action = registerCategoryFacet({
       ...defaultCategoryFacetOptions,
-      facetSearch: {...defaultFacetSearchOptions},
       ...options,
       facetId,
     });