+++ packages/headless/src/test/mock-insight-state.ts
@@ -4,7 +4,10 @@ import {getInsightConfigurationInitialState} from '../features/insight-configura
 import {getInsightInterfaceInitialState} from '../features/insight-interface/insight-interface-state';
 import {getInsightCaseContextSearchInitialState} from '../features/insight-search/insight-case-context-state';
 import {getPaginationInitialState} from '../features/pagination/pagination-state';
+import {getQuerySetInitialState} from '../features/query-set/query-set-state';
+import {getQuerySuggestSetInitialState} from '../features/query-suggest/query-suggest-state';
 import {getQueryInitialState} from '../features/query/query-state';
+import {getResultPreviewInitialState} from '../features/result-preview/result-preview-state';
 import {getSearchHubInitialState} from '../features/search-hub/search-hub-state';
 import {getSearchInitialState} from '../features/search/search-state';
 import {InsightAppState} from '../state/insight-app-state';
@@ -23,6 +26,9 @@ export function buildMockInsightState(
     query: getQueryInitialState(),
     facetSet: getFacetSetInitialState(),
     pagination: getPaginationInitialState(),
+    resultPreview: getResultPreviewInitialState(),
+    querySet: getQuerySetInitialState(),
+    querySuggest: getQuerySuggestSetInitialState(),
     ...config,
   };
 }