+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,95 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.25.12](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.11...@coveo/quantic@1.25.12) (2022-06-21)
+
+
+### Bug Fixes
+
+* **headless:** update recent queries only when query has results ([#2110](https://github.com/coveo/ui-kit/issues/2110)) ([623674b](https://github.com/coveo/ui-kit/commit/623674b6d64d0f03a129651fdfa859b570baff54))
+
+
+
+
+
+## [1.25.11](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.10...@coveo/quantic@1.25.11) (2022-06-21)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.10](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.9...@coveo/quantic@1.25.10) (2022-06-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.9](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.8...@coveo/quantic@1.25.9) (2022-06-20)
+
+
+### Bug Fixes
+
+* **quantic:** automatically  closing the tab dropdown when it loses focus ([#2114](https://github.com/coveo/ui-kit/issues/2114)) ([0488be6](https://github.com/coveo/ui-kit/commit/0488be63cb0981494eddfc9940ac77c976ab24b2))
+
+
+
+
+
+## [1.25.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.7...@coveo/quantic@1.25.8) (2022-06-20)
+
+
+### Bug Fixes
+
+* **quantic:** display clear all filters button only when there is active filters ([#2106](https://github.com/coveo/ui-kit/issues/2106)) ([9aabbf9](https://github.com/coveo/ui-kit/commit/9aabbf953c0803749f4cd34cda6385fa29ee6f16))
+
+
+
+
+
+## [1.25.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.6...@coveo/quantic@1.25.7) (2022-06-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.6](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.5...@coveo/quantic@1.25.6) (2022-06-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.4...@coveo/quantic@1.25.5) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.3...@coveo/quantic@1.25.4) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.2...@coveo/quantic@1.25.3) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.25.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.1...@coveo/quantic@1.25.2) (2022-06-16)
 
 **Note:** Version bump only for package @coveo/quantic